# backend/api/openapi.gen.json
{
  "components": {
    "headers": {
...
      "Redirect": {
        "description": "Where the browser goes next",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "SetOidcCookie": {
        "description": "The login's state cookie `__Host-cowork-oidc`: `HttpOnly; Secure;\nSameSite=Lax; Path=/; Max-Age=600`. It holds the `state`, the `nonce`, the\nPKCE verifier, `return_to` and the time, sealed with AES-256-GCM under a key\nderived from `COWORK_SESSION_KEY` (docs/adr/0029 D1); the callback reads it\nonce and clears it.\n",
        "schema": {
          "type": "string"
        }
      }
    },
...
    "responses": {
      "Problem": {
        "content": {
          "application/problem+json": {
            "example": {
              "code": "not_found",
              "detail": "no such tenant",
              "instance": "/api/v1/tenants/acme",
              "request_id": "0199a3c2-1d2e-7f00-8000-0000000000ff",
              "status": 404,
              "title": "Not found",
              "type": "https://cowork.dev/problems/not-found"
            },
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "An RFC 9457 problem details body with a stable `code`, the `request_id`\nand, for validation failures, `errors[]` (docs/adr/0047).\n",
        "headers": {
          "X-Request-Id": {
            "description": "The request's id, also in the body and in the log line",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "schemas": {
...
      "FieldError": {
        "properties": {
          "current": {
            "description": "The server's current value of the field, on a 412; null when the field is empty",
            "type": [
              "string",
              "number",
              "integer",
              "boolean",
              "object",
              "array",
              "null"
            ],
            "x-go-type": "interface{}"
          },
          "message": {
            "type": "string"
          },
          "pointer": {
            "description": "A JSON pointer into the body, or query:\u003cname\u003e / header:\u003cname\u003e",
            "type": "string"
          }
        },
        "required": [
          "pointer",
          "message"
        ],
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProblemCode"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "instance": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "type": "object"
      },
      "ProblemCode": {
        "description": "The stable code of a problem (docs/adr/0047 D2)",
        "enum": [
          "validation_failed",
          "idempotency_key_required",
          "invalid_cursor",
          "page_too_deep",
          "unauthenticated",
          "token_expired",
          "token_revoked",
          "not_allowed",
          "invalid_credentials",
          "forbidden",
          "insufficient_scope",
          "agent_forbidden",
          "session_required",
          "password_change_required",
          "not_initialised",
          "csrf",
          "not_found",
          "person_not_found",
          "method_not_allowed",
          "username_taken",
          "tenant_slug_taken",
          "project_key_taken",
          "person_ambiguous",
          "grant_exists",
          "mapping_exists",
          "last_admin",
          "project_archived",
          "state_conflict",
          "parent_cycle",
          "link_cycle",
          "open_prerequisites",
          "period_locked",
          "attachment_limit",
          "uploads_disabled",
          "precondition_failed",
          "payload_too_large",
          "unsupported_media_type",
          "idempotency_mismatch",
          "precondition_required",
          "too_many_attempts",
          "internal",
          "not_ready",
          "timeout",
          "backend_unreachable"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearerToken": {
        "description": "A personal access token, `cwk_` and 43 base62 characters\n(docs/adr/0035 D1). `X-Cowork-Agent: name/model/session` marks the\nrequest as an agent's; a token with the agent flag is an agent's\nwith or without it, and no header removes the flag (docs/adr/0036).\nA request that carries an `Authorization` header is a token's, and\nits cookie, if any, is not looked at. A token of a person the identity\nprovider's groups no longer admit is refused with `401 not_allowed`,\nchecked against the person's groups as of their last login or groups\nrefresh, at most `COWORK_OIDC_GROUPS_REFRESH` apart (docs/adr/0035 D8).\n",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "The session of a browser login: an opaque id of 256 random bits, set\nby `POST /auth/local` or `GET /auth/callback` as `HttpOnly; Secure;\nSameSite=Lax; Path=/` (docs/adr/0031 D2). A session acts as its\nperson with no agent flag and with the whole of the person's role: it\nhas no scope of its own.\nIts writes are held to the CSRF check of docs/adr/0037 D1: `Origin`\n(or `Referer`) equal to the installation's `COWORK_BASE_URL` and\n`X-Requested-With: cowork`.\n",
        "in": "cookie",
        "name": "__Host-cowork-session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "A multi-tenant backlog for people and the agents that work for them.\n\nEvery route that names a person is authenticated with one of two\ncredentials. A personal access token, `Authorization: Bearer cwk_…`\n(docs/adr/0035), is for scripts and agents; the session cookie of a\nbrowser login (docs/adr/0031 — `POST /auth/local`, or the identity\nprovider's `GET /auth/callback`, docs/adr/0029) is for the UI. Each\noperation says which it takes: most take either, a few — creating a token,\na tenant or a local account, resetting or changing a password, logging out,\nand the administration acts that give access: adding a member, setting a\ngrant, making or changing a group mapping, restricting or opening a\nproject, putting a person on its access list — a session only. A request\nof a session that writes must come from the installation's origin and\ncarry `X-Requested-With: cowork` (docs/adr/0037). A request marked as an\nagent's — by the token's agent flag, or by an `X-Cowork-Agent:\nname/model/session` header — is bound by the agent rules\n(docs/adr/0036, 0043). Errors are RFC 9457 problem details with a stable\n`code` (docs/adr/0047).\n",
    "title": "cowork",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/callback": {
      "get": {
        "description": "Where the issuer sends the browser back: `COWORK_BASE_URL` +\n`/auth/callback`, the redirect URI registered with the identity\nprovider (docs/adr/0029 D4). The UI never calls it.\n\nThe state cookie of `GET /auth/oidc/login` must be present, younger than\nten minutes, and hold the `state` the issuer returned. The code is\nexchanged with the PKCE verifier; the ID token's signature, issuer,\naudience, expiry and nonce are verified (D1); the groups are read from\n`COWORK_OIDC_GROUPS_CLAIM` — in the ID token, else from UserInfo (D2) —\nand must pass the gate: one of them is in `COWORK_OIDC_ALLOWED_GROUPS` or\nis `COWORK_ADMIN_GROUP` (docs/adr/0030 D1). The person is found by\nissuer and subject, or made (docs/adr/0029 D5), and their memberships\nare derived from the tenants' group mappings (docs/adr/0030 D2). Then the\nsession is made as the local login makes it — a new cookie, and the\nsession the request presented ended (docs/adr/0031 D5) — the state\ncookie is cleared, and the answer is `303` to `return_to`.\n\nEvery failure is `303` to `/login?error=\u003ccode\u003e\u0026return=\u003creturn_to\u003e`, the\nstate cookie cleared and no session made. `return` is the path the login\nbegan with, for the page's next attempt: it is there when the state\ncookie still opens, stale or not, and the path is not `/`, the page's\ndefault. The codes: `oidc_failed` — the state cookie is missing, stale\nor does not match, the exchange or the verification failed, or the\nissuer answered with `error`; `not_allowed` — the person is outside the\ngate, or deactivated (docs/adr/0024 D5); `not_initialised` — no tenant\nexists and the person is not in `COWORK_ADMIN_GROUP` (docs/adr/0032 D5);\n`oidc_unavailable` — no identity provider is configured. A refused\nlogin is recorded (docs/adr/0026); the reason reaches the log, never the\npage, and neither ever holds a code or a token.\n\nThe issuer may add parameters this document does not name, such as `iss`\n(RFC 9207) or `session_state`: the route takes them and reads none of\nthem (`x-cowork-open-query`).\n",
        "operationId": "oidcCallback",
        "parameters": [
          {
            "description": "The authorization code",
            "in": "query",
            "name": "code",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The `state` of the login, as the issuer returns it",
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The issuer's error, sent instead of `code` (RFC 6749 4.1.2.1)",
            "in": "query",
            "name": "error",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "error_description",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "To `return_to`, logged in, or to `/login?error=\u003ccode\u003e\u0026return=\u003creturn_to\u003e`",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Redirect"
              },
              "Set-Cookie": {
                "description": "Two cookies: the state cookie `__Host-cowork-oidc` cleared with\n`Max-Age=0`, and after a login the session cookie\n`__Host-cowork-session` (docs/adr/0031 D2).\n",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [],
        "summary": "The identity provider's return",
        "tags": [
          "auth"
        ],
        "x-cowork-open-query": true
      }
    },
...
    "/auth/oidc/login": {
      "get": {
        "description": "A navigation of the browser — the login page's button sets\n`window.location` here — and never a call: the answer is a redirect\n(docs/adr/0029 D1). With an identity provider configured it is `302` to\nthe issuer's authorization endpoint, for the authorization code flow\nwith PKCE (`S256`), a `state` and a `nonce`, and it sets the cookie\n`__Host-cowork-oidc` that holds them with `return_to` for the callback.\nWithout one, or while its gate admits nobody — the cases in which\n`GET /auth/options` answers `oidc: false` — it is `303` to\n`/login?error=oidc_unavailable\u0026return=\u003creturn_to\u003e`.\n\n`return_to` is where the browser lands after the login: a path of this\ninstallation that starts with `/`, not with `//` or `/\\`, holds no\ncontrol character and is at most 2048 characters long. Anything else,\nor none, is `/`; it is never refused, so a bad link still leads to a\nlogin.\n\nNo credential: whoever comes here has none yet. Nothing is decided here\nthat a forged link could abuse — the session is made by the callback,\nand only for the browser that holds the state cookie this answer set.\n",
        "operationId": "loginOidc",
        "parameters": [
          {
            "description": "The path to land on after the login, and the `return` of a refusal's redirect to the login page; anything that is not a path of this installation is `/`",
            "in": "query",
            "name": "return_to",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "To the issuer's authorization endpoint; the state cookie is set",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Redirect"
              },
              "Set-Cookie": {
                "$ref": "#/components/headers/SetOidcCookie"
              }
            }
          },
          "303": {
            "description": "To `/login?error=oidc_unavailable\u0026return=\u003creturn_to\u003e` — no identity provider is configured, or its gate admits nobody",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Redirect"
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [],
        "summary": "Start a login through the identity provider",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerToken": []
    },
    {
      "sessionCookie": []
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ],
...
}